create serialize submodule for block
